Module cplfs_api::controller

Implementation of a physical disk and a very simple device controller for it. The device and its contents are represented by a file in your file system, that is memory mapped and stored in a Device struct. When initializing the controller, you have to provide it with either a path to a non-existing file, which will then be created and used as the contents of your device, or to an existing file, which will be opened and the contents of which will be checked. Provides a basic block read and write operation on a device at a given offset. The memory-mapped file is what the read and write functions operate on.

EXTRA: Note that this explicit block-level abstraction is not required for a file system at this level of abstraction, but added it to make our model a more realistic representation of a real-life file system. No provisions have been made to properly lock and unlock the file that is used to back the file system, so do not fiddle with it while a file system is running, as this leads to undefined behavior. (e.g. the fs2 crate could be used to explicitly implement locking, if so desired)

Structs

Device
Struct representing the state of a hard drive disk (HDD). The implementation of this structure is the controller that allows us to read disk blocks from the disk, and write disk blocks to the disk.

Enums

DiskState
Small enum, used to specify whether we expect to open a new file system