Struct cplfs_api::controller::Device

pub struct Device {
    pub block_size: u64,
    pub nblocks: u64,
    // some fields omitted
}

Struct representing the state of a hard drive disk (HDD). The implementation of this structure is the controller that allows us to read disk blocks from the disk, and write disk blocks to the disk.

EXTRA: As a side note, it would be nicer if we could make both the Device and the Block polymorphic in block_size i.e. add <block_size : u64> and write the signature of e.g. the read_block function as: read_block(&self, index: u64) -> anyhow::Result<Block<block_size>> with self now of type Device<block_size> This would statically enforce blocks to have the right size when they are read or written. This type of genericism, which can be seen as a lightweight version of dependent types, is also referred to as const generics. No const generics possible in stable rust yet, to parameterize over block_size. This sort of functionality can be achieved using templates in C++, and is also available in Rust's nightly build, but as of yet not in the stable fraction of Rust.

Fields
block_size: u64

Size of the blocks that this disk reads and writes

nblocks: u64

Total number of blocks this disk consists of

Implementations

impl Device

pub fn create_device<P: AsRef<Path>>(
path: P,
block_size: u64,
nblocks: u64,
ds: DiskState
) -> Result<Device>

Core function to that handles both new and load, based on the value of the switch ds, representing whether we want to load or create a disk

pub fn new<P: AsRef<Path>>(
path: P,
block_size: u64,
nblocks: u64
) -> Result<Device>

Create a new disk device, given:

A path to store its image

A block_size to define the size of each unit to be read or written, in bytes

The total number of blocks in the disk This new device will have contents 0 at each address.

Note that if block_size is smaller than the size of the main types (for the super block, inodes, etc.) used in this assignment, the file system will crash at runtime. We did not program our code defensively to cope with this. This function will return an error, if the file represented by path already exists.

pub fn load<P: AsRef<Path>>(
path: P,
block_size: u64,
nblocks: u64
) -> Result<Device>

Load an existing disk device, given its block_size and the number of blocks its file system ought to contain. This function will return an error, if the file represented by path does not yet exist.

pub fn destruct(self)

End the lifetime of this disk, and remove the file backing it on disk Assumes that you have not made any other links to the backing file Panics if removing the file fails

pub fn device_size(&self) -> u64

Size of this device in bytes

pub fn device_path(&self) -> &Path

Path of the file backing this device

pub fn read_block(&self, index: u64) -> Result<Block>

Read the block with index index from the device Results in an error if the block index is too high The block is returned in the form of a Block structure

pub fn write_block(&mut self, b: &Block) -> Result<()>

Write a given block buf into the device at index index Fails if buf is not exactly block-sized, or if the provided index is too high