Trait cplfs_api::fs::BlockSupport

pub trait BlockSupport: FileSysSupport {
    fn b_get(&self, i: u64) -> Result<Block, Self::Error>;

    fn b_put(&mut self, b: &Block) -> Result<(), Self::Error>;

    fn b_free(&mut self, i: u64) -> Result<(), Self::Error>;

    fn b_zero(&mut self, i: u64) -> Result<(), Self::Error>;

    fn b_alloc(&mut self) -> Result<u64, Self::Error>;

    fn sup_get(&self) -> Result<SuperBlock, Self::Error>;

    fn sup_put(&mut self, sup: &SuperBlock) -> Result<(), Self::Error>;
}

This trait adds block-level operations to your file system Do not forget to make sure that mkfs now correctly initializes all inodes, as described above.

Required methods

fn b_get(&self, i: u64) -> Result<Block, Self::Error>

Read the nth block of the entire disk and return it
The implementation of this method should be trivial

fn b_put(&mut self, b: &Block) -> Result<(), Self::Error>

Write the nth block of the entire disk and return it
The implementation of this method should be trivial

fn b_free(&mut self, i: u64) -> Result<(), Self::Error>

Free the ith block in the block data region, by setting the ith bit in the free bit map region to zero. To disambiguate the binary representation, we define the order of the bits within each byte you read from right to left, i.e. the byte '0b0000_0001' has its first bit set to one, i.e. has its bit with index 0 set to 1. Be careful not to change any bits for other blocks around the ith block by overwriting an entire byte in the free bit map region! Consider writing some auxiliary methods for more fine-grained access to Blocks, by wrapping the provided block type in a different type.

Remember that you cannot access memory on a per-bit basis! The bitwise operators & (bitwise AND), | (bitwise OR), ^ (bitwise XOR) exist; use those, together with the bit-shift operators >> and <<, to change the correct bit (or use an external, stable crate).

If the ith block is already a free block, the state of memory should not change, and this method should return an error Also errors if the index i is out of bounds, i.e. if it is higher than the number of data blocks

fn b_zero(&mut self, i: u64) -> Result<(), Self::Error>

Zeroes the ith block in the block data region in memory (without freeing it)
Errors if the index i is out of bounds, i.e. if it is higher than the number of data blocks

fn b_alloc(&mut self) -> Result<u64, Self::Error>

Allocate the first free block, starting from the beginning, in the data block region, thereby setting its bit in the bitmap region to one and the entire contents of the block to zero. Again, you will need bit-wise operators to implement this function. Obviously, only blocks that have not been allocated yet, i.e. do not have their bit set, can still be allocated. Returns the index (within the data region) of the newly allocated block. Make sure to load each bitmap block only once in your implementation. Errors appropriately if no blocks are available.

fn sup_get(&self) -> Result<SuperBlock, Self::Error>

Get the superblock describing the current file system

fn sup_put(&mut self, sup: &SuperBlock) -> Result<(), Self::Error>

Write the superblock to this file system's underlying device (and cache it, depending on how you implement your file system) Note that in general, when you write to a block, you read the block first and then overwrite the parts you need to change before writing it back, to make sure all other parts of the block remain unchanged. In this case, that is not strictly necessary, as the superblock is the only useful thing that is stored on the first disk block. However, in case other data were to be stored past the superblock struct in the future, do implement this function in this conservative way.

Loading content...