Trait cplfs_api::fs::DirectorySupport

pub trait DirectorySupport: InodeSupport {
    fn new_de(inum: u64, name: &str) -> Option<DirEntry>;

    fn get_name_str(de: &DirEntry) -> String;

    fn set_name_str(de: &mut DirEntry, name: &str) -> Option<()>;

    fn dirlookup(
        &self, 
        inode: &Self::Inode, 
        name: &str
    ) -> Result<(Self::Inode, u64), Self::Error>;

    fn dirlink(
        &mut self, 
        inode: &mut Self::Inode, 
        name: &str, 
        inum: u64
    ) -> Result<u64, Self::Error>;
}

This trait adds the abstraction of directories and their entries to the file system Additionally, it supports some convenience methods that allow you to use directory entries with string names (the reason these methods are defined here and not in a trait, is to avoid forcing you to wrap the DirEntry type in another type of your own). Do not forget to make sure that mkfs now defines a valid (currently empty) root directory.

Note that i_free is possibly unsafe for directories; if the freed directory is the last one to point to some entry and we free the directory's inode, the entry will be left dangling. To combat this, system calls will never call i_free directly, but rather call wrapper methods such as e.g. unlink (see the PathSupport trait and optional assignment for more information), to check all the necessary preconditions. The standard directory entries of ".." and ".", that are present in any directory but the parent directory, will be implemented as part of the PathSupport extension too.

Required methods

fn new_de(inum: u64, name: &str) -> Option<DirEntry>

Create a new directory entry, given inum and name Returns None if an invalid name is provided Uses set_name_str to set the directory entry name

fn get_name_str(de: &DirEntry) -> String

Get the name of this directory entry as a String Loop over the characters in the direntry's name until you encounter a final character \0 OR until you reach the end of the character array, and return the concatenation of these characters as the Direntry's name.

fn set_name_str(de: &mut DirEntry, name: &str) -> Option<()>

Set the name of this directory entry to the given name, if the given name is valid, i.e. it is

non-empty

consists of alphanumeric characters only, or is equal to "." or ".."

is sufficiently short when converted to characters If the name is shorter than DIRNAME_SIZE, insert a '\0' at the end so you can still correctly read it after. Returns None in case of an invalid name

fn dirlookup(
&self,
inode: &Self::Inode,
name: &str
) -> Result<(Self::Inode, u64), Self::Error>

Look for a directory entry named name in a given inode, representing a directory. Make sure that the inode you are calling this function with is up to date wrt. the one on disk! If found, return the inode corresponding to this directory entry, and the byte offset (from the start of the inode contents) it was found at, as a pair. Only inspects directory entries that fall within the size of the given inode. Errors if the given inode is not of directory type. If the entry is not found, return a suitable error that you can handle later, in the implementation of dirlink below.

NOTE: self is mutably borrowed here because might want to call iget

fn dirlink(
&mut self,
inode: &mut Self::Inode,
name: &str,
inum: u64
) -> Result<u64, Self::Error>

Write a new directory entry with contents name and inum into the directory represented by inode. Make sure that the inode you are calling this function with is up to date wrt. the one on disk! In case the directory has no free entries, append a new entry to the end and increase the size of inode. This might require allocating a new block as well. In case you implemented the optional InodeRWSupport assignment already , you could use the i_write method here.

When a place for the given inode is found, looks up the inode corresponding to inum and increase its nlink field with 1 on disk (unless inum and inode's number are equal, then nothing happens, as this is a self-reference). Errors if

name is invalid, or is already an entry inside inode.

inode is not a directory.

errors and does nothing if the inode corresponding to inum is not currently in use.

Returns the byte offset at which the entry was written into the given inode

EXTRA:In our model, we do not have to worry about hardlink-loops in our file system tree, as we currently provide no way of duplicating inodes.

Loading content...