Trait cplfs_api::fs::FileSysSupport

pub trait FileSysSupport: Sized {
    type Error: Error;
    fn sb_valid(sb: &SuperBlock) -> bool;

    fn mkfs<P: AsRef<Path>>(
        path: P, 
        sb: &SuperBlock
    ) -> Result<Self, Self::Error>;

    fn mountfs(dev: Device) -> Result<Self, Self::Error>;

    fn unmountfs(self) -> Device;
}

General trait that each filesystem should implement, that allows us to set up, tear down and load file systems in the tests Additionally, this trait also defines the error type that is used in all of the other traits (which will require implementing this trait) Be warned that the implementation of this trait cannot be kept the same throughout the assignment! Depending on the abstractions your current file system is aware of, it might impose more strict conditions when loading a file system, and might perform extra steps when defining one. For example, when implementing the block layer for your file system, it should not create a root inode as part of mkfs, and neither should it check for one when mounting an existing file system. These steps should only be implemented at higher levels of abstraction. In other words, you should not have one blanket implementation of this trait for your entire project, and it will be beneficial to you when e.g. defining the inode abstraction, to take your solution from the block layer, including its implementation of this trait, and create a wrapper around it that (re)implements this and any other relevant traits, so that you can delegate calls to this trait to the wrapped object, and define extra steps on top of them.

Consult the documentation of SuperBlock for more detailed information on the proper layout of the file system and the meaning of the different superblock components

HINT: consider caching your filesystem superblock when creating or loading a file system, so you do not have to read it each time you need it

Associated Types

type Error: Error

The type of the errors of your implementation. Read more about how to define your own errors in the documentation of error_given.

Note the “supertrait” your error type must implement: error::Error This is the base trait for all Rust errors, click on it to see the methods you must implement.

Note that you can use the same error type for multiple parts of the assignment.

Loading content...
Required methods

fn sb_valid(sb: &SuperBlock) -> bool

Static method to check if a given superblock represents a valid file system. You will need this both when creating a new file system, and loading an existing one from disk

More concretely, the following conditions need to hold to make a SuperBlock a possibly valid representation of a file system

the regions have to appear in the right order

the regions have to be sufficiently large to hold ninodes inodes (calculated using the size of your inodes) and to hold and keep track of ndatablocks datablocks

the regions have to physically fit on the disk together, i.e. fall within the first nblocks blocks

fn mkfs<P: AsRef<Path>>(path: P, sb: &SuperBlock) -> Result<Self, Self::Error>

Method to create and mount a new file system from scratch, given a super block (which is a bit more convenient to work with than a bunch of parameters) and a path. What exactly this method does, depends on the level of abstraction you are implementing.

This method always does the following, regardless of the layer of abstraction:

Check if the given superblock is a valid file system superblock

Create a new Device at the given path, to allow the file system to communicate with it

Then, subdivide the freshly created device image into the previously described regions:

A super block containing the file system metadata at block index 0

This is only relevant for the inode layer and up, i.e. ignore this set-up step in the disk block layer. A list of blocks containing inodes, where all inodes as marked as "free" (since they are free, the rest of their contents is irrelevant and hence unspecified).
You only need to do the following once you support directories: The first inode (with number 1) describes the "root path", i.e. the path "/" on UNIX systems. It is initially empty, i.e. has no directory entries inside it. The root node has its nlink field set to 1 from the start, even though there are no references to it, so that it cannot be deallocated. Note that inodes start counting at one, and NOT at zero, to avoid confusion with the error return value 0 in the kernel. However, to avoid off-by-1 errors, room for inode 0 is still allocated in the first inode block on disk (i.e. inode 1 is not stored at address 0 of this block). This space will under normal circumstances never be used.

A bitmap keeping track of the occupied memory blocks. This bitmap should initially mark all blocks as "free", as no block allocations have happened.

A data region to contain the memory blocks themselves Since all blocks are marked as free after initialization and allocating a block should set its contents to 0, the contents of this region is unimportant. This data region is assumed to run until the end of the file system.

Think about whether the initial Device you create inside this method, which will contain 0 at each address (see the documentation), is a valid disk representation, or if it requires extra initialization for one or more of the above regions. (Hint for the inode layer and up: watch out for the inodes; an all-0 inode will not necessarily come out well during deserialization, and probably needs to be overwritten by an actually free inode)

Make sure your underlying device is in a consistent state and matches the above enumeration at the end of this function.

IMPORTANT NOTE: In case you need to loop over inodes here or anywhere else in this project, do so efficiently, i.e. if you need to read/write multiple inodes in the same block, only load and store this block once!

EXTRA: mkfs is inspired by the unix command of the same name (although this version also immediately mounts the file system)

fn mountfs(dev: Device) -> Result<Self, Self::Error>

Given an existing Device called dev, make sure that its image corresponds to a valid file system by reading its superblock and checking the following conditions:

The superblock is a valid superblock

The block size and number of blocks of the device and superblock agree

If these conditions are satisfied, wrap the given Device in a file system and return it.

You do not need to deserialize each individual object in each region to check that it is indeed a valid object; to keep matters simple, we will assume that the contents of each region has been properly initialized. Additionally, we could add dev to the return type to reclaim ownership in case of an error, but we do not bother recovering invalid devices, for simplicity reasons.

fn unmountfs(self) -> Device

Unmount the give file system, thereby consuming it Returns the image of the file system, i.e. the Device backing it. The implementation of this method should be almost trivial

Loading content...