Trait cplfs_api::fs::InodeCacheSupport

pub trait InodeCacheSupport: InodeSupport {
    fn i_get_mut(&mut self, i: u64) -> Result<Self::Inode, Self::Error>;

    fn is_cached(&self, inum: u64) -> bool;

    fn mkfs_cached<P: AsRef<Path>>(
        path: P, 
        sb: &SuperBlock, 
        nb_cache_entries: u64
    ) -> Result<Self, Self::Error>;

    fn mountfs_cached(
        dev: Device, 
        nb_cache_entries: u64
    ) -> Result<Self, Self::Error>;
}

Support caching for inodes. Read more about what exactly this entails in assignment g_caching_inodes.rs in the solution folder.

Required methods

fn i_get_mut(&mut self, i: u64) -> Result<Self::Inode, Self::Error>

i_get_mut is a new version of i_get that takes a mutable reference to self. This is required since getting an inode from the disk potentially alters the cache. The original implementation of i_get should now only read values from the cache, and not fetch them from disk, so that self does not have to be mutable there This function then subsumes the behavior of iget, by behaving in the following way;

Finds the inode corresponding to i and returns it. Looks in the cache first and returns a copy if the inode is present. If the inode is not there, reads it from disk and puts it in the cache. When looking for an entry to replace in the cache, i_get_mut will evict the first entry it finds that currently has no other references to it (you can check this using the strong_count method of the Rc type). Before evicting the old entry, make sure to persist its contents to the disk. After eviction, the read DInode is wrapped in a CachedInode, written into the cache, and a copy of it is returned to the caller.

fn is_cached(&self, inum: u64) -> bool

Is the given inode currently in the inode cache?

fn mkfs_cached<P: AsRef<Path>>(
path: P,
sb: &SuperBlock,
nb_cache_entries: u64
) -> Result<Self, Self::Error>

Alternative version of mkfs, that allows us to specify the number of entries in the inode cache. Interpret the original mfks function as a more specific variant of this function, where the number of cache entries for inodes is fixed to 5.

fn mountfs_cached(
dev: Device,
nb_cache_entries: u64
) -> Result<Self, Self::Error>

Alternative version of mountfs, that allows us to specify the number of entries in the inode cache. Interpret the original mountfs function as a more specific variant of this function, where the number of cache entries for inodes is fixed to 5.

Loading content...