Trait cplfs_api::fs::InodeRWSupport

pub trait InodeRWSupport: InodeSupport {
    fn i_read(
        &self, 
        inode: &Self::Inode, 
        buf: &mut Buffer, 
        off: u64, 
        n: u64
    ) -> Result<u64, Self::Error>;

    fn i_write(
        &mut self, 
        inode: &mut Self::Inode, 
        buf: &Buffer, 
        off: u64, 
        n: u64
    ) -> Result<(), Self::Error>;
}

This trait additionally provides support to read and write from inodes using buffers; the data structure that we used before to hold the contents of a Block. The reason we use Buffers objects and not simply the raw data, is because Buffers also support some methods relating to (de)serialization, which might come in handy in some situations. We ignore arithmetic over- and underflow in the following functions, for reasons of simplicity.

Required methods

fn i_read(
&self,
inode: &Self::Inode,
buf: &mut Buffer,
off: u64,
n: u64
) -> Result<u64, Self::Error>

Read n bytes of data from inode inode into the given buffer buf, starting from byte offset off in the inode inode Returns the amount of bytes read, so that clients of this function can check if something went wrong If the end of the file is reached while reading, stop reading. If a read starts at inode.get_size(), returns with 0 bytes read. By contrast, returns an error and does not read anything in case the provided index falls further outside of the file's bounds. If buf cannot hold n bytes of data, reads until buf is full instead.

fn i_write(
&mut self,
inode: &mut Self::Inode,
buf: &Buffer,
off: u64,
n: u64
) -> Result<(), Self::Error>

Write n bytes of data from the given buffer buff into the inode inode, starting from byte offset off If the end of the file is reached while writing, continue writing. If necessary, start allocating extra blocks to expand the file and continue writing into the new blocks. However, returns 0 and does not write anything in case the provided offset falls outside of the file's bounds. If the inode changes while writing, do not forget to write it back to the disk too. Returns an error if buf cannot hold at least n bytes of data. If the write would make the inode exceed its maximum possible size, do nothing and return an error.

Loading content...