Trait cplfs_api::fs::InodeSupport

pub trait InodeSupport: BlockSupport {
    type Inode: InodeLike;
    fn i_get(&self, i: u64) -> Result<Self::Inode, Self::Error>;

    fn i_put(&mut self, ino: &Self::Inode) -> Result<(), Self::Error>;

    fn i_free(&mut self, i: u64) -> Result<(), Self::Error>;

    fn i_alloc(&mut self, ft: FType) -> Result<u64, Self::Error>;

    fn i_trunc(&mut self, inode: &mut Self::Inode) -> Result<(), Self::Error>;
}

This trait adds the abstraction of inodes to your file system. Do not forget that mkfs should now result in a filesystem where each inode is marked as free. Note that this trait does not yet support ways of growing inodes. Growing inodes will become possible when you implement the read and write calls on them in one of the optional assignments.

Associated Types

type Inode: InodeLike

The type you use for your inodes at runtime. Initially, you can probably get away with the inode type I provided. In the extensions, we can replace this type with a different type of inode that provides more flexibility (e.g. aliasing behavior, and a variable number of blocks)

Loading content...
Required methods

fn i_get(&self, i: u64) -> Result<Self::Inode, Self::Error>

Read the disk inode with index i and wrap it into an inode object. Should error if i is greater than the number of inodes in the system.

fn i_put(&mut self, ino: &Self::Inode) -> Result<(), Self::Error>

Write the given inode back to the disk at the correct position

fn i_free(&mut self, i: u64) -> Result<(), Self::Error>

Similar to b_free, but now for inodes instead of data blocks. A big difference is that this method should only free an inode if it is no longer referenced anywhere else in the file system, i.e. if its nlink field is equal to zero. In this case, the $i$th inode in the inode region is freed by setting its ft field to TFree. Additionally, all valid blocks belonging to this file, listed in its direct_blocks array, have to be deallocated (and set to address 0) as well. In case nlink is not equal to zero, this method does nothing. Returns an error if i does not correspond to an inode number, or if the inode is already free.

fn i_alloc(&mut self, ft: FType) -> Result<u64, Self::Error>

Similar to b_alloc, but now for inodes instead of data Allocates the first free dinode (i.e. lowest inum) it comes across, and sets (on disk):

this inode's FType to ft

this inode's size and nlink fields to 0, as it currently has no blocks and is not referenced in the file system The inode with index 0 should never be allocated. Errors appropriately if no inodes are available Only read each inode block once in your implementation

fn i_trunc(&mut self, inode: &mut Self::Inode) -> Result<(), Self::Error>

Truncate the given inode, i.e. release its contents (without freeing it). Sets all of the given inode's direct_blocks to point to address 0. Releases all blocks direct_blocks belonging to this inode, and sets its size to 0. Changes both the given inode and the corresponding inode on the disk. Note that only the valid blocks should be released as only these are allocated. In other words, do not blindly release all values listed in the direct_blocks field

Loading content...