Trait cplfs_api::fs::PathSupport

pub trait PathSupport: DirectorySupport {
    fn valid_path(path: &str) -> bool;

    fn get_cwd(&self) -> String;

    fn set_cwd(&mut self, path: &str) -> Option<()>;

    fn resolve_path(&self, path: &str) -> Result<Self::Inode, Self::Error>;

    fn mkdir(&mut self, path: &str) -> Result<Self::Inode, Self::Error>;

    fn unlink(&mut self, path: &str) -> Result<(), Self::Error>;
}

Enhance the previous directory support with a notion of file paths (both absolute and relative), enabling the following:

Allows looking up inodes along file paths, to allow for easier navigation.

Allows creating and unlinking (i.e. removing) directories at a given path location. The previously implemented i_alloc and i_free suffice for regular files, but can result in inconsistencies for directories.

Keeps track of the current working directory in your file system (pick an appropriate type to do this! **WARNING: give this some thought **). Usually, each process has its own current working directory, but since we only model a single process, we just store the current working directory in the file system.

Do not forget to adapt the call to mkfs, so that the root directory is no longer empty, but like all other directories, gets initialized with "." and ".." as its first two entries. The root directory is a special case, however, in the sense that ".." also points back to the root directory itself, since it has no parent. Make sure the nlink field of the root directory is equal to 1 still, after the initialization phase. Additionally, both the mkfs and mountfs calls initialize the current working directory to the root directory.

IMPORTANT for this assignment;

We assume that none of the directories involved in the cwd's path will be deleted while we are in this directory. We do not check this at any point, nor do we lock any of those files or increase their nlink; we just assume that it will not happen. Additionally, we do not check that the cwd actually also exists in the file system; this is the responsibility of our clients

In general, it is not necessarily the case the parent of a directory is the previous directory in the path! For example, starting from /test then following the path ../alternative by reading the file system, does not necessarily have us end up in /alternative. The reason is that the parent of test might be some different directory entirely, because the dirlink method below allows us to re-register inodes in different directories. However, you can assume that the parent of the cwd is the previous directory in the path when appending a relative path in the set_cwd method. This is also how terminals usually operate when e.g. following symbolic links and then running cd .., in order not to confuse users. That being said, any paths provided to other methods than set_cwd should go through the file system, i.e. read ".." at each point to figure out what the actual parent inode is. If this explanation is unclear to you, consult the tests provided with the assignment

Required methods

fn valid_path(path: &str) -> bool

Returns true iff the given string represents a valid path We support two different path formats:

absolute paths start with the character "/" and specify a path starting from the root node

relative paths start with "." or ".." and denote respectively a path starting from the current working directory, and a path starting from the parent of the current working directory

A path is valid iff:

It is not empty

It is either absolute or relative in form

It consists of a "/"-separated sequence of names

It does not end in a "/" (the only exception being the path "/" itself)

Each one of these names is a valid directory entry name (see the DirectorySupport trait)

Instead of an alphanumeric name, the special entries "." and ".." can also appear anywhere in the file path. Their meaning is then similar to their meaning they have at the start of a relative path; "." specifies staying in the current directory, whereas ".." specifies moving up to the parent directory.

fn get_cwd(&self) -> String

Return the current working directory as a String path The value you return here should be an absolute path with respect to the root path /, and should not contain the special names "." or ".."

fn set_cwd(&mut self, path: &str) -> Option<()>

Set the current working directory to the path provided in the given String, if it is a valid path. Returns none if the provided path was invalid.

If a relative path is provided, then that path is interpreted with respect to the current working directory. As stated above, you should then not go through the file system to determine the new cwd, but rather have ".." cancel out the previous directory name, i.e. if the cwd is test/child then providing ../child_new to this method, will unconditionally result in a new cwd of test/child_new, whereas it will not necessarily lead there when we read through the file system in the other methods.

fn resolve_path(&self, path: &str) -> Result<Self::Inode, Self::Error>

Given a path name (possibly relative to the cwd), look up the inode corresponding to this path (the final inode could be either a file or a directory), and return it. Works as follows:

Figures out what inode the cwd corresponds to (skip this step if path is absolute)

Reads path through the file system

Returns an error if

the path is invalid

any directory, referenced in the path (or the cwd), does not exist

any of the intermediate names refers to a file that is not of directory type.

fn mkdir(&mut self, path: &str) -> Result<Self::Inode, Self::Error>

Create a new directory at the given path, where the last name of the path is the name for the new directory. For example, the path /test/dir will create a directory named dir in the parent directory test. Returns the newly created inode in case of success. Note that the newly created directory is already referenced once, i.e. its nlink field is not 0.

New directories are created with 2 default entries in them, i.e. "." and ".." (in this order), that point to the current directory's inode and its parent's inode, respectively. This allows us to easily go back to the parent directory of the current directory, and to succinctly reference sibling files and folders in the same directory. Note that the addition of ".." causes the nlink field of the parent directory to increase by 1.

Error

if the path is not valid

if the path's prefix (i.e. the part of the path without the name of the directory that we are about to create) does not exist in the file system yet.

if the last part of the path is not a valid directory name (i.e. it cannot be "." or "..")

fn unlink(&mut self, path: &str) -> Result<(), Self::Error>

Remove the directory entry located at path path, i.e. set the inum of this entry to 0 and the name of the entry to all zeroes, i.e. to "0".repeat(*DIRNAME_SIZE) For example, the path /test/dir will (on success) delete the directory entry "dir" in the parent directory test. In essence, the inverse of mkdir.

Decreases the nlink field of the deleted entry by 1 (unless we just deleted a cyclic reference to the parent), and frees the inode in case nlink decreases to 0. In the latter case, a reference to the inode also gets deleted (since ".." in the entry's inode disappears when it is freed)

Errors and does nothing else in the following cases:

the path ends in "." or ".."; these entries cannot be removed

the path is not valid

the path does not exist yet in the file system, which includes the next bullet;

the entry is not present in the directory

the entry we are about to delete is itself a directory and non-empty (apart from the 2 default entries) - note: you cannot judge emptiness just from the size of the file, as it might contain directory entries that were previously unlinked as well

Loading content...