Struct cplfs_api::types::Buffer

pub struct Buffer { /* fields omitted */ }

Buffer abstraction, representing some data on the heap. Buffers can have any size, and will be used further on to build file system Blocks with, but also as output to read and write functions on files Support regular read and write methods, but also (de)serialization of structures implementing the appropriate traits

Implementations

impl Buffer

pub fn new(data: Box<[u8]>) -> Buffer

Create a new buffer, having the given data slice as its data

pub fn new_zero(len: u64) -> Buffer

Create an all-zero buffer, with contents length of len

pub fn len(&self) -> u64

Size of the underlying block data

pub fn contents_as_ref(&self) -> &[u8]

Return a reference to this block's contents

pub fn read_data(&self, data: &mut [u8], offset: u64) -> Result<()>

Reads data from the given buffer into the data buffer, starting at the given offset. Returns the number of bytes that were read, or an error in case of failure. If the function does not return an error, the number of bytes read should always be equal to data.len().

pub fn write_data(&mut self, data: &[u8], offset: u64) -> Result<()>

Writes data from the given slice into the data. If the function does not return an error, the number of bytes written should always be equal to data.len()`.

pub fn deserialize_from<S>(&self, offset: u64) -> Result<S> where
S: DeserializeOwned,

Read any object that implements the DeserializeOwned trait from this buffer

EXTRA: Note that since this method takes ownership of the deserialized data, the link with the original data in the block necessarily breaks. This is not what you would have in a high-performance C implementation, as you would simply perform a cast of the part of memory you are interested in to a struct, without having to worry about lifetimes. To keep things simple and not have additional lifetime dependencies and unsafe code here, this method was not implemented as such.

pub fn serialize_into<S>(&mut self, stru: &S, offset: u64) -> Result<()> where
S: Serialize,

Write any object that implements the Serialize trait into this buffer Goes through write_data so that the appropriate error get triggered. Alternatively, we could go through serialize_into bincode and use the standard error.