Struct cplfs_api::types::Inode

pub struct Inode {
    pub inum: u64,
    pub disk_node: DInode,
}

Wrapper around disk inodes DInode used for in-memory inodes. Additionally contains the number of the inode inum. This information is not required as long as the inode is stored on disk, as it is implicit from the block in which the inode is stored. This is analogous to a Block explicitly keeping track of its block number

Fields
inum: u64

inode number

disk_node: DInode

the disk contents corresponding to inum

Implementations

impl Inode

pub fn new(inum: u64, disk_node: DInode) -> Inode

Create a new inode