Struct cplfs_api::types::SuperBlock

pub struct SuperBlock {
    pub block_size: u64,
    pub nblocks: u64,
    pub ninodes: u64,
    pub inodestart: u64,
    pub ndatablocks: u64,
    pub bmapstart: u64,
    pub datastart: u64,
}

Structure representing all file system metadata that we are interested in, and hence the file system's structure. Note that the size of the Superblock struct does not necessarily have to be a full block, as it can just be read from disk contiguously. Rather, the size of SuperBlock must be at most as large as a single disk block. Derives Serialize and Deserialize so we can easily write this block to the disk and read it again after.

The layout of the simple file system model we use is as follows: [super block | inode blocks | free bit map | data blocks] , where each component has the following meaning:

super block: aggregates all the file system meta-data including the sizes of all subsequent regions. This is the first block that is read by the file system driver when loading an existing file system, and the first block to be written by the driver in case a new file system is initialized. This area should consist of a single block, i.e. the SuperBlock type defined below should not take up more space in memory than a single block, defined by Disk.block_size in controller.rs.

inode blocks: a sequence of blocks containing all the inode metadata. This region contains all inodes in order, starting from inode 1 (the root directory, i.e. the directory on your computer with path ""), all the way up to the last inode. The number of inodes stored in each block is equal to the floor of the block size divided by the inode size, i.e. blocks are packed with inodes, and individual inodes are always entirely stored in a single block (they are never broken up over multiple blocks).

free bit map: a sequence of blocks keeping track of the allocation state (allocated or free) of all disk blocks in the next data block region. The nth bit in this sequence specifies whether or not the nth data block is currently in use.

data blocks: contain the actual file and directory data, as a long sequence of disk blocks.

EXTRA: Since we do not support logging, there is no need for an additional memory region to store any logs in Also note that in contrast to more realistic device layouts, we ignore the fact that the first block of the device is often reserved for bootstrapping code, and makes use of e.g. a Master Boot Record (MBR) or Volume Boot Record (VBR). EXTRA: Note that just like blocks, inodes are not being cached either. The consequence is that the users of our APIs are responsible for ensuring that they aren't handling different aliases to the same inode without realizing it. This will not scale well to a parallellized setting. In our case, this is no major problem, as we have no parallellism, and we have simple system call interactions, that will not handle a lot of inodes at the same time, and will hence not need to perform many of those inode equality checks.

Fields
block_size: u64

Size of the blocks in the current file system, in BYTES
In a real world application, this block size does not necessarily match the size of the sectors on the device itself, but for simplicity reasons we assume this value and the disk sector size in Device.block_size to always be equal

nblocks: u64

Number of blocks in the entire file system, including this block and the 3 other file system regions
This number does not necessarily have to equal the size of the disk this file system is stored on, as long as the file system fits on the disk

ninodes: u64

Number of inodes that we keep track of in the inode region
This number does not necessarily have to fill up the entire region, i.e. it is possible to make the inode region unnecessarily big

inodestart: u64

The block index of the first block of inodes
Since the super block is only a single block long and located at index 0, this will usually be the block with index
The inode region runs until bmapstart
The inode region is assumed to be sufficiently long to contain niondes inodes

ndatablocks: u64

Number of data blocks that we keep track of in the bitmap region
This number does not necessarily have to fill up the entire bitmap or data region, i.e. it is possible to make the bitmap and/or block data region unnecessarily big

bmapstart: u64

The block index of the first block of the free bit map region
The free bit map region runs until datastart
The free bit map region is assumed to be at least ndatablocks bits large\

datastart: u64

The block index of the first block of the data blocks region
The data block region runs until nblocks, i.e. the end of the file system
The data block region is assumed to be at least ndatablocks blocks large