Trait cplfs_api::types::InodeLike

pub trait InodeLike: Sized {
    fn new(
        inum: u64, 
        ft: &FType, 
        nlink: u64, 
        size: u64, 
        blocks: &[u64]
    ) -> Option<Self>;

    fn get_ft(&self) -> FType;

    fn get_nlink(&self) -> u64;

    fn get_size(&self) -> u64;

    fn get_block(&self, i: u64) -> u64;

    fn get_inum(&self) -> u64;
}

Trait for inode-like behavior, so that we can have different inodes later on, without having to change the interfaces of the inode trait Solely used in testing, so does not require setter methods

Required methods

fn new(
inum: u64,
ft: &FType,
nlink: u64,
size: u64,
blocks: &[u64]
) -> Option<Self>

Create a new inode from the given parameters The user of this method is responsible for making sure that the parameters are consistent with each other and the rest of the code

fn get_ft(&self) -> FType

Get the file type of this inode

fn get_nlink(&self) -> u64

Get the number of links to this inode in the file system

fn get_size(&self) -> u64

Get the size of this inode in bytes

fn get_block(&self, i: u64) -> u64

Get the address of the ith block pointed to by this inode, if there is any block under index i Note that this function's behavior is undefined for unallocated indexes i (as these are commonly set to 0). It is up to the caller to figure out whether the index i was sensible, based on get_size() and the external file system block size.

fn get_inum(&self) -> u64

Get the number of this inode on the disk

Loading content...